Flow of viscous fluid through a converging nozzle
It is known that for viscous/inviscid incompressible fluid flows, the velocity will show an inversely proportional relationship to the X-section the fluid is passing through. This investigation shows the increase in velocity corresponding to the decrease in X-sectional area. High pressure region is situated where the X-section is larger(before the throat). On outlet (also on inlet)velocity drops as we reach closer to the wall, suggesting viscosity(graph is attached). For inviscid fluid this characteristic will not be present. Besides the pressure before the throat for inviscid fluid will be lower than of viscous as more pressure is required to move a viscous fluid through a pipe/nozzle. The inviscid fluid will follow Bernoulli's Equation(Though i simulated only the viscous one :(    ).As always I provided the mesh file (though designing this 2d mesh is very easy)